+++ router.php
@@ -1,8 +1,19 @@
 <?php
 // htfaker - router.php
 
+namespace Attogram\Htfaker;
+
+use Symfony\Component\HttpFoundation\Request;
+
 require 'vendor/autoload.php';
 
-$htfaker = new \Attogram\Htfaker\Htfaker(true);
+$htfaker = new \Attogram\Htfaker\Htfaker(
+  Request::createFromGlobals(),
+  true // debug
+);
 
-return $htfaker->run();
+// "If this script returns FALSE, then the requested resource is returned as-is.
+// Otherwise the script's output is returned to the browser."
+if( !$htfaker->run() ) {
+    return false;
+}