+++ webtests/index.php
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>webtests</title>
+  </head>
+  <body>
+    webtests
+  </body>
+</html>